+++ packages/headless/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.1](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.7.0...@coveo/headless@1.7.1) (2021-07-15)
+
+**Note:** Version bump only for package @coveo/headless
+
+
+
+
+
 # [1.7.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.6.3...@coveo/headless@1.7.0) (2021-07-14)