+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.35.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.34.2...@coveo/quantic@0.35.0) (2021-07-15)
+
+
+### Features
+
+* **quantic:** fix-facets-selected-state ([#964](https://github.com/coveo/ui-kit/issues/964)) ([8012578](https://github.com/coveo/ui-kit/commit/8012578ae105bbd0c2384e8947d80c4c2e29e057))
+
+
+
+
+
+## [0.34.2](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.34.1...@coveo/quantic@0.34.2) (2021-07-15)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
 ## [0.34.1](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.34.0...@coveo/quantic@0.34.1) (2021-07-14)
 
 **Note:** Version bump only for package @coveo/quantic